feat: replace OpenAI embeddings with sentence transformers local model. refactor to move tinyvector and embeddings model clients to app state

+++ src/types/mod.rs
@@ -31,7 +31,7 @@ pub struct Document {
 pub struct Embedding {
     pub source_id: String,
     pub doc_path: String,
-    pub chunk: u32,
+    pub chunk: usize, // TODO rename to index
     pub blob: String,
     pub vector: Vec<f32>,
 }
